Modifying main constructor of `Easymore`
The main constructor has turned into a more Pythonic way of defining
one, as well as assigning `Easymore` as the main class of the `easymore`
package.

Necessary docstrings are added to `Easymore` in numpydoc style, however,
the definition of Attributes and useful Methods are left empty for
future development process.

Reported-by: Kasra Keshavarz <kasra.keshavarz1@ucalgary.ca>