Reverting from using tempfile module
As job submission might take a while to run, the files produced using
the tempfile library may get deleted by the OS, therefore, the temporary
file has been switched to be created inside the temporary directory
given by the user.

Furthermore, this measure, will assure the workflow is saved for future
reference. Polishing up the CLI and structure of it is warranted to
assure it can be used by the larger community.